# perf(utils): linear scan `is_uri`