@juntuu 2023-08-23
Hi, just bumped into this while browsing.

I'm not at all familiar with the project, so can't say if the performance improvement would outweigh the complexity cost, and what would be the effect in real workloads.

However, if I'm not mistaken this change introduces two differences in the matching:

1. does not require a `:` in uri (`is_uri("hello") == true`)
2. the windows "drive letter" can be more than one letter (`is_uri("hello:\\") == false`)

Because the function is checking for uri and not windows path, I think 2. is more correct now. Uri should not contain `\` anyway.

But 1. seems like a bug.

I think the following (pseudocode) should fix 1. while keeping 2.:
```lua
if filename[1] not in [a-zA-Z] then
  return false
end
for i = 2, #filename do
  if filename[i] not in [a-zA-Z0-9.+-] then
    return false
  elseif filename[i] == ':' then
    return filename[i+1] ~= '\\'
  end
end
return false
```