@max397574 2023-08-23
@juntuu that pseudocode definitely won't work since there is no chance the `elseif` statement ever is true
it would be true if the character is a colon but if that's the case it will already return false because of the if statement