@juntuu 2023-08-23
> that pseudocode won't work

Oops, good catch. Thanks!

I flipped the conditions now, so the `:` is checked first.