@jamestrew 2023-08-26
Good catches to both!
I guess under ideal conditions, something like `is_uri('hello')` or really anything that's not a full path shouldn't happen but for the sake of correctness, I like the suggested change.

No significant performance difference between my original implementation and the suggested change.

```
$ hyperfine 'luajit before_linux.lua' 'luajit after_linux.lua' --warmup 10 --runs 100
Benchmark 1: luajit before_linux.lua
  Time (mean ± σ):      27.2 ms ±   0.5 ms    [User: 22.6 ms, System: 4.5 ms]
  Range (min … max):    26.4 ms …  29.9 ms    100 runs

Benchmark 2: luajit after_linux.lua
  Time (mean ± σ):      27.5 ms ±   0.8 ms    [User: 22.8 ms, System: 4.7 ms]
  Range (min … max):    26.0 ms …  30.5 ms    100 runs

Summary
  luajit before_linux.lua ran
    1.01 ± 0.03 times faster than luajit after_linux.lua
```

```
$ hyperfine 'luajit before_win.lua' 'luajit after_win.lua' --warmup 10 --runs 100
Benchmark 1: luajit before_win.lua
  Time (mean ± σ):      26.9 ms ±   0.5 ms    [User: 23.6 ms, System: 3.6 ms]
  Range (min … max):    26.0 ms …  28.8 ms    100 runs

Benchmark 2: luajit after_win.lua
  Time (mean ± σ):      25.9 ms ±   0.7 ms    [User: 22.9 ms, System: 3.3 ms]
  Range (min … max):    24.8 ms …  28.9 ms    100 runs

Summary
  luajit after_win.lua ran
    1.04 ± 0.03 times faster than luajit before_win.lua
```